requirements.txt.txt
asgiref==3.4.1
Django==4.0
django-heroku==0.3.1
psycopg2==2.9.3
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.13.0
gunicorn==20.1.0
Markdown==3.3.6
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
whitenoise==5.3.0